Bug 1736365 - Allow 64-bit offsets in m64 instructions.  r=yury
The main strategy employed here is to treat offsets > UINT32_MAX as an
outlier and to cause those to force the computation of the full
effective address early in the compilation pipelines.  The rest of the
pipelines thus deal only with 32-bit offsets and remain unchanged.

Testing here is probably a little light, though it is meaningful.
More test cases will land before this is enabled in Nightly.

Differential Revision: https://phabricator.services.mozilla.com/D128984

+++ js/src/jit/MIR.h
@@ -9044,10 +9044,10 @@ class MWasmBoundsCheck : public MBinaryInstruction, public NoTypePolicy::Data {
 };
 
 class MWasmAddOffset : public MUnaryInstruction, public NoTypePolicy::Data {
-  uint32_t offset_;
+  uint64_t offset_;
   wasm::BytecodeOffset bytecodeOffset_;
 
-  MWasmAddOffset(MDefinition* base, uint32_t offset,
+  MWasmAddOffset(MDefinition* base, uint64_t offset,
                  wasm::BytecodeOffset bytecodeOffset)
       : MUnaryInstruction(classOpcode, base),
         offset_(offset),
@@ -9067,7 +9067,7 @@ class MWasmAddOffset : public MUnaryInstruction, public NoTypePolicy::Data {
 
   AliasSet getAliasSet() const override { return AliasSet::None(); }
 
-  uint32_t offset() const { return offset_; }
+  uint64_t offset() const { return offset_; }
   wasm::BytecodeOffset bytecodeOffset() const { return bytecodeOffset_; }
 };
